create logistics data source
