+++ graphql/types/Profile.graphql
@@ -72,7 +72,6 @@ type Address {
   addressId: String
   cacheId: ID
   id: String
-  addressId: String
   userId: String
   receiverName: String
   complement: String
