+++ node/resolvers/httpResolver.ts
@@ -23,7 +23,7 @@ export interface HttpResolverOptions {
   merge?: ResponseMerger
 }
 
-export type HttpResolver<T> = (root: any, args: any, context: ServiceContext) => Promise<T>
+export type HttpResolver<T> = (root: any, args: any, context: Context) => Promise<T>
 
 export default <T=any>(options: HttpResolverOptions): HttpResolver<T> => {
   return async (root, args, { vtex: ioContext, request: { headers: { cookie, 'x-forwarded-host': host } }, response }: Context) => {
